+++ src/Pages/Validator/Components/ValidationTxs.js
@@ -30,7 +30,7 @@ const ValidationTxs = () => {
     { displayName: 'Fee', dataName: 'fee' },
     { displayName: 'Signer', dataName: 'signers' },
     { displayName: 'Status', dataName: 'status' },
-    { displayName: 'Timestamp', dataName: 'timestamp' },
+    { displayName: 'Timestamp', dataName: 'time' },
   ];
 
   return (