+++ src/Pages/Validator/Components/ValidatorDelegations.js
@@ -21,7 +21,7 @@ const ValidatorDelegations = () => {
 
   // Table header values in order
   const tableHeaders = [
-    { displayName: 'Address', dataName: 'address' },
+    { displayName: 'Address', dataName: 'delegatorAddr' },
     { displayName: 'Amount', dataName: 'amount' },
     { displayName: 'Shares', dataName: 'shares' },
   ];